Update i18n.js
Added option missingBehaviour, which avoids to have to create strings when developing or simple strings that do not need a declaration in the default language.
Combine that with missingTranslationPrefix and you can still identify which strings are translated and which are guessed.

This commit is based on pr#220 with some small changes added..

Updated against latest upstream/master